Use datetime objects in the filter API instead of unix timestamps
In preparation for the new categorization algorithm which will work on
local datetime objects.

Note: in some tests now using a fixed date instead of the current
time. Some of the tests fail if the tested time period contains a
daylight saving time switch -- with using local times, DST becomes an
issue. I'm ignoring it for now though, will clarify in later commits.