Simplify calculation of rejected objects
This simplifies the filtering loops so that they just determine the
items to be accepted. The rejected items are calculated in a final
single step as the (set-) difference of the original and the accepted
items.

This improves clarity (the code expresses clearly that everything that
isn't accepted is rejected) and will make it easier to change the
filtering algorithm in the following commits.

As a tradeoff the filtering might become slower (there is now an
additional list copy and the calculation of the difference), but I
think the improved clarity is worth it.