Simplify (overlapping) rule evaluation
This changes the way rules are applied to a given set of items
as follows:

- for each category and timecount, the most recent item is accepted.

  (except for the "recent" category, where, as previously, the
  "timecount" most recent items are accepted)

- rules are applied independently; the set of accepted items for a set
  of rules is the union of accepted item sets for the individual rules

The main motivation for this change is that it simplifies verifying
the results, both manually and with programmatic tests: to verify that
a set of rules was correctly applied you only have to check the expected
accepted items for each single rule.

Previously this wasn't possible for overlapping categories and
timecounts (see previous "overlapping_rules_accept_additional_items"
test).

This changes the results of overlapping rules (see e.g. the
10-days-2-weeks test), so that the ...overlap tests in
TestTimeFilterMass break. I'm removing them for now, also because they
contained a lot of reasoning about the previous behavior which is now
obsolete. Tests demonstrating the new behavior will be added later in
the branch.