Demonstrate motivation for changing overlapping rules behavior
The results of overlapping rules can be complicated to explain and
understand (see for example the comment on the "test_10_days_2_weeks"
test in test_api.py).

This adds a minimal test demonstrating (hopefully) the source of
this complication: rules aren't applied independently but interact.
To understand the results of a combination of rules, one must also
understand that interaction.

To allow an easier transition to the new filtering algorithm, I'd like
to reduce that complexity in the following commit.