Cleanup mass tests: use a single item list with distributed moddates
This replaces the fsegen item generator used in TestFilterMass.
fsegen would generate many items, but most of them  with identical
time, and there were large gaps between items of different times.
The new make_dense_fses function tries to create a more realistic
distribution of items with small gaps between them and no identical
times.

Also this removes the fse9 item list -- all of the TestFilterMass
tests check that the maximum possible number of items is found for
each rule, so it is ok to use the larger list for all tests.