Switch to calendar based filtering
With this, time differences aren't calculated linearly (2 days ago ==
48 hours ago == 48 * 3600 seconds ago) but based on local calendar
dates and times (1 month ago == "some time last month", 2 days ago ==
"any time on the day before yesterday" etc.)

This fixes a fundemental problem of the filtering algorithm: for each
time category, it dispatched items in "buckets" relative to the given
reference time. For each bucket covered by the filtering rules, it
accepted the newest item and rejected all other items in the bucket.

However, that idea to only keep the newest item of each bucket seems
to conflict with the varying distribution of items between buckets --
for every filter run, the newest item of a bucket may be different,
buckets populated before may become empty, and, most importantly,
items that were alone in a bucket before may be combined with a newer
item in that item's bucket.

This could cause unexpected deletion of items. For example, if the
reference time was 10:30 two items with modtimes of 9:20 and 8:25
would be put into different "hourly" buckets -- bucket 1 and 2,
respectively. At reference time 11:20, both items would be put in
bucket 2. Assuming a rule of "hours:x" (x>=3), the item at 8:25 would
be deleted at 11:20 but retained at 11:30 -- then the items would end
up in bucket 2 and 3.

With this change, bucket boundaries are "fixed" on the calendar. While
the "ordinal" number of a bucket (e.g. whether it's 1, 2 or 3 hours
ago) depends on the reference time, the distribution of items between
buckets stays stable. After a bucket has been filtered so that it only
contains one item, further deletions only occur if the bucket as a whole
isn't covered by a rule any more, or if newer items are added to the
bucket between filter runs; but not by just re-running the filter at a
slightly different time.